Refactor application form
